requirements.txt
Cython
matplotlib
numpy
omegaconf>=2.0.1
onnx
onnx-simplifier
opencv-python
pyaml
pycocotools
pytorch-lightning>=1.9.0,<2.0.0
tabulate
tensorboard
termcolor
#torch>=1.10,<2.0
torchmetrics
#torchvision
tqdm
deeplake
